cleanup prototype transforms functional tests
